** fixed more errors and warnings
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@5649 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/FORMAT/Base64.C
@@ -57,7 +57,7 @@ namespace OpenMS
 			str.push_back('\0');
 		}
 
-		QByteArray original = QByteArray::fromRawData(str.c_str(), str.size());
+		QByteArray original = QByteArray::fromRawData(str.c_str(),(int) str.size());
 		QByteArray base64_compressed;
 		if (zlib_compression)
 		{
